+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 ### Fixed
 - Deployment scripts
 - Database introspection method
+- Deleted two duplicated volume entries
 
 ## 2022-03-03 -- v0.0.1
 ### Added
 - Initial commit of dockerized CCE/CCR Search API
+- ON DELETE CASCADE settings for Volume entries and associated child entries