+++ CHANGELOG.md
@@ -1,13 +1,19 @@
 # CHANGELOG
 
+## 2022-03-25 -- v0.0.3
+### Fixed
+- Deleted two duplicated volume entries
+
+### Added
+- ON DELETE CASCADE settings for Volume entries and associated child entries
+- Added delete-orphan settings back to cce model
+
 ## 2022-03-09 -- v0.0.2
 ### Fixed
 - Deployment scripts
 - Database introspection method
-- Deleted two duplicated volume entries
 
 ## 2022-03-03 -- v0.0.1
 ### Added
 - Initial commit of dockerized CCE/CCR Search API
-- ON DELETE CASCADE settings for Volume entries and associated child entries
-- Added delete-orphan settings back to cce model
+